requirements.txt
discord.py==1.3.2
Adafruit_Blinka==4.2.0
discord==1.0.1
python-dotenv==0.12.0
